convert to mysql

+++ database/migrations/2019_11_12_071707_author_post.php
@@ -14,8 +14,8 @@ class AuthorPost extends Migration
     public function up()
     {
         Schema::create('post_user', function (Blueprint $table) {
-            $table->integer('user_id')->unsigned();
-            $table->integer('post_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
+            $table->bigInteger('post_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->foreign('post_id')->references('id')->on('posts')->onDelete('cascade');
         });
