+++ database/migrations/2019_11_13_204909_create_schedule_events_table.php
@@ -17,7 +17,7 @@ class CreateScheduleEventsTable extends Migration
             $table->bigIncrements('id');
             $table->text('title');
             $table->date('date');
-            $table->unsignedInteger('schedule_id');
+            $table->unsignedBigInteger('schedule_id');
             $table->foreign('schedule_id')->references('id')->on('schedules')->onDelete('cascade');
             $table->timestamps();
         });
